remove unnecessary helpers for setting labels and annotations
