--- pkg/util/doc.go
@@ -1,4 +1,0 @@
-// Package util implements various utility functions used in both testing and
-// implementation of OpenShift. Package util may not depend on any other
-// package in the OpenShift package tree.
-package util
